+++ tests/specs/outline-split.test.ts
@@ -25,3 +25,4 @@ import '../suites/outline-split/v094StructuredOutlineDocument.suite';
 import '../suites/outline-split/v094StructuredOutlineEditor.suite';
 import '../suites/outline-split/v094StructuredOutlineVisual.suite';
 import '../suites/outline-split/v099OutlineGuideGeometry.suite';
+import '../suites/outline-split/v0910OutlineGuidesAndReveal.suite';